Move MetabrowseOptions methods to Metabrowse
