+++ modules/cli/src/main/scala/scala/cli/commands/Metabrowse.scala
@@ -10,6 +10,7 @@ import scala.build.{Build, BuildThreads, Logger}
 import scala.cli.CurrentParams
 import scala.cli.commands.util.SharedOptionsUtil._
 import scala.cli.packaging.Library
+import scala.util.Properties
 
 object Metabrowse extends ScalaCommand[MetabrowseOptions] {
   override def hidden     = true
@@ -22,6 +23,21 @@ object Metabrowse extends ScalaCommand[MetabrowseOptions] {
 
   override def sharedOptions(options: MetabrowseOptions) = Some(options.shared)
 
+  private def metabrowseBinaryUrl(
+    scalaVersion: String,
+    options: MetabrowseOptions
+  ): (String, Boolean) = {
+    import options._
+    val osArchSuffix0 = osArchSuffix.map(_.trim).filter(_.nonEmpty)
+      .getOrElse(FetchExternalBinary.platformSuffix(supportsMusl = false))
+    val metabrowseTag0           = metabrowseTag.getOrElse("latest")
+    val metabrowseGithubOrgName0 = metabrowseGithubOrgName.getOrElse("alexarchambault/metabrowse")
+    val metabrowseExtension0     = if (Properties.isWin) ".zip" else ".gz"
+    val url =
+      s"https://github.com/$metabrowseGithubOrgName0/releases/download/$metabrowseTag0/metabrowse-$scalaVersion-$osArchSuffix0$metabrowseExtension0"
+    (url, !metabrowseTag0.startsWith("v"))
+  }
+
   def run(options: MetabrowseOptions, args: RemainingArgs): Unit = {
     CurrentParams.verbosity = options.shared.logging.verbosity
     val inputs = options.shared.inputsOrExit(args)
@@ -29,8 +45,16 @@ object Metabrowse extends ScalaCommand[MetabrowseOptions] {
 
     val logger = options.shared.logger
 
-    val initialBuildOptions = options.buildOptions
-    val threads             = BuildThreads.create()
+    val baseOptions = options.shared.buildOptions(enableJmh = false, jmhVersion = None)
+    val initialBuildOptions = baseOptions.copy(
+      classPathOptions = baseOptions.classPathOptions.copy(
+        fetchSources = Some(true)
+      ),
+      javaOptions = baseOptions.javaOptions.copy(
+        jvmIdOpt = baseOptions.javaOptions.jvmIdOpt.orElse(Some("8"))
+      )
+    )
+    val threads = BuildThreads.create()
 
     val compilerMaker = options.shared.compilerMaker(threads)
 
@@ -73,7 +97,7 @@ object Metabrowse extends ScalaCommand[MetabrowseOptions] {
       .map(os.Path(_, os.pwd))
       .getOrElse {
         val (url, changing) =
-          options.metabrowseBinaryUrl(successfulBuild.scalaParams.scalaVersion)
+          metabrowseBinaryUrl(successfulBuild.scalaParams.scalaVersion, options)
         FetchExternalBinary.fetch(
           url,
           changing,
