+++ modules/cli/src/main/scala/scala/cli/commands/MetabrowseOptions.scala
@@ -2,11 +2,6 @@ package scala.cli.commands
 
 import caseapp._
 
-import scala.build.internal.FetchExternalBinary
-import scala.build.options.BuildOptions
-import scala.cli.commands.util.SharedOptionsUtil._
-import scala.util.Properties
-
 // format: off
 @HelpMessage("Browse Scala code and its dependencies in the browser")
 final case class MetabrowseOptions(
@@ -39,29 +34,6 @@ final case class MetabrowseOptions(
     metabrowseDialect: Option[String] = None
 ) {
   // format: on
-
-  def metabrowseBinaryUrl(scalaVersion: String): (String, Boolean) = {
-    val osArchSuffix0 = osArchSuffix.map(_.trim).filter(_.nonEmpty)
-      .getOrElse(FetchExternalBinary.platformSuffix(supportsMusl = false))
-    val metabrowseTag0           = metabrowseTag.getOrElse("latest")
-    val metabrowseGithubOrgName0 = metabrowseGithubOrgName.getOrElse("alexarchambault/metabrowse")
-    val metabrowseExtension0     = if (Properties.isWin) ".zip" else ".gz"
-    val url =
-      s"https://github.com/$metabrowseGithubOrgName0/releases/download/$metabrowseTag0/metabrowse-$scalaVersion-$osArchSuffix0$metabrowseExtension0"
-    (url, !metabrowseTag0.startsWith("v"))
-  }
-
-  def buildOptions: BuildOptions = {
-    val baseOptions = shared.buildOptions(enableJmh = false, jmhVersion = None)
-    baseOptions.copy(
-      classPathOptions = baseOptions.classPathOptions.copy(
-        fetchSources = Some(true)
-      ),
-      javaOptions = baseOptions.javaOptions.copy(
-        jvmIdOpt = baseOptions.javaOptions.jvmIdOpt.orElse(Some("8"))
-      )
-    )
-  }
 }
 
 object MetabrowseOptions {
